Added the packaging page to the seller panel
Sellers can now record the boxes and cartons they pack into, under Settings →
Fulfillment beside stock locations and delivery methods. The operator's page
moved into the shared package so both panels render the same one, and the
operator's list gained a column saying who owns each row.

The variant editor's carton picker starts working for sellers as a
consequence: it already read packaging through the panel's client, which had
nothing registered until now.
