+++ packages/dashboard-core/src/index.ts
@@ -62,6 +62,7 @@ export * from './hooks/use-export'
 export * from './hooks/use-global-search'
 export * from './hooks/use-host-form'
 export * from './hooks/use-import'
+export * from './hooks/use-package-types'
 export * from './hooks/use-resource-mutation'
 export * from './hooks/use-stock-locations'
 export * from './hooks/use-switch-admin-locale'
@@ -104,6 +105,7 @@ export * from './lib/slot-registry'
 export * from './lib/table-registry'
 export { ensureTimestampColumns } from './lib/timestamp-columns'
 export * from './lib/type-labels'
+export * from './pages/package-types'
 export * from './pages/stock-locations'
 // ---------------------------------------------------------------------------
 // Plugin facade — re-exported for convenience; same API as `/plugin` subpath
@@ -128,4 +130,5 @@ export * from './providers/permission-provider'
 export * from './providers/sticky-header-provider'
 export * from './providers/store-provider'
 export * from './providers/tenant-provider'
+export * from './schemas/package-type'
 export * from './schemas/stock-location'
