+++ packages/dashboard-core/src/locales/fr.json
@@ -624,6 +624,45 @@
       "name": {
         "label": "Nom"
       },
+      "package_type": {
+        "name": {
+          "label": "Nom"
+        },
+        "kind": {
+          "label": "Type",
+          "help": "Seul un carton maître peut être l'emballage d'un produit."
+        },
+        "length": {
+          "label": "Longueur"
+        },
+        "width": {
+          "label": "Largeur"
+        },
+        "height": {
+          "label": "Hauteur"
+        },
+        "dimensions_unit": {
+          "label": "Unité"
+        },
+        "dimensions": {
+          "help": "Utilisées pour le poids volumétrique et pour les mètres cubes des paliers de fret."
+        },
+        "weight": {
+          "label": "Poids à vide",
+          "help": "L'emballage lui-même, ajouté à son contenu."
+        },
+        "max_weight": {
+          "label": "Poids max.",
+          "help": "Ce qu'il peut porter."
+        },
+        "weight_unit": {
+          "label": "Unité"
+        },
+        "default": {
+          "label": "Carton par défaut",
+          "help": "La base de chaque cotation colis. Un seul par boutique."
+        }
+      },
       "number": {
         "label": "Numéro"
       },
@@ -989,12 +1028,46 @@
       "tax_category_saved": "Catégorie de taxe enregistrée",
       "updated": "Mis à jour"
     },
+    "package_types": {
+      "table_title": "Types d'emballage",
+      "search_placeholder": "Rechercher un emballage",
+      "empty": "Aucun emballage. Ajoutez votre carton d'expédition, ainsi que les cartons et palettes des commandes en gros.",
+      "add_cta": "Ajouter un type d'emballage",
+      "default_badge": "Par défaut",
+      "edit_sheet_title": "Type d'emballage",
+      "create_description": "Un carton, une palette ou un conteneur dans lequel cette boutique emballe ses produits.",
+      "edit_description": "Les produits emballés dans ce carton suivent ses dimensions : une modification les corrige tous.",
+      "dimensions_column": "Dimensions",
+      "volume_column": "Volume",
+      "kinds": {
+        "box": "Carton",
+        "envelope": "Enveloppe",
+        "carton": "Carton maître",
+        "pallet": "Palette",
+        "container": "Conteneur"
+      },
+      "messages": {
+        "added": "Type d'emballage ajouté",
+        "updated": "Type d'emballage mis à jour",
+        "deleted": "Type d'emballage supprimé"
+      },
+      "delete_confirm": {
+        "title": "Supprimer ce type d'emballage ?",
+        "message": "{{name}} sera supprimé. Les produits qui y sont emballés doivent d'abord être reconditionnés."
+      },
+      "validation": {
+        "non_negative": "doit être supérieur ou égal à zéro"
+      }
+    },
     "pages": {
       "settings": {
         "stock_locations": {
           "add_sheet_title": "Ajouter un emplacement de stock",
           "edit_sheet_title": "Modifier l'emplacement de stock",
           "search_placeholder": "Rechercher des emplacements de stock…"
+        },
+        "package_types": {
+          "add_sheet_title": "Ajouter un type d'emballage"
         }
       },
       "products": {
