+++ packages/dashboard-core/src/locales/pl.json
@@ -625,6 +625,45 @@
       "name": {
         "label": "Nazwa"
       },
+      "package_type": {
+        "name": {
+          "label": "Nazwa"
+        },
+        "kind": {
+          "label": "Rodzaj",
+          "help": "Tylko karton zbiorczy może być opakowaniem produktu."
+        },
+        "length": {
+          "label": "Długość"
+        },
+        "width": {
+          "label": "Szerokość"
+        },
+        "height": {
+          "label": "Wysokość"
+        },
+        "dimensions_unit": {
+          "label": "Jednostka"
+        },
+        "dimensions": {
+          "help": "Do wagi objętościowej przewoźnika i metrów sześciennych progów frachtowych."
+        },
+        "weight": {
+          "label": "Waga pustego",
+          "help": "Samo opakowanie, doliczane do zawartości."
+        },
+        "max_weight": {
+          "label": "Maks. waga",
+          "help": "Ile może unieść."
+        },
+        "weight_unit": {
+          "label": "Jednostka"
+        },
+        "default": {
+          "label": "Domyślny karton sklepu",
+          "help": "Podstawa każdej wyceny paczki. Tylko jeden na sklep."
+        }
+      },
       "number": {
         "label": "Numer"
       },
@@ -990,12 +1029,46 @@
       "tax_category_saved": "Zapisano kategorię podatkową",
       "updated": "Zaktualizowano"
     },
+    "package_types": {
+      "table_title": "Rodzaje opakowań",
+      "search_placeholder": "Szukaj opakowania",
+      "empty": "Brak opakowań. Dodaj karton wysyłkowy oraz kartony i palety dla zamówień hurtowych.",
+      "add_cta": "Dodaj rodzaj opakowania",
+      "default_badge": "Domyślne sklepu",
+      "edit_sheet_title": "Rodzaj opakowania",
+      "create_description": "Karton, paleta lub kontener, w który ten sklep pakuje towar.",
+      "edit_description": "Produkty pakowane w ten karton przyjmują jego wymiary — jedna zmiana poprawia wszystkie.",
+      "dimensions_column": "Wymiary",
+      "volume_column": "Objętość",
+      "kinds": {
+        "box": "Karton",
+        "envelope": "Koperta",
+        "carton": "Karton zbiorczy",
+        "pallet": "Paleta",
+        "container": "Kontener"
+      },
+      "messages": {
+        "added": "Dodano rodzaj opakowania",
+        "updated": "Zaktualizowano rodzaj opakowania",
+        "deleted": "Usunięto rodzaj opakowania"
+      },
+      "delete_confirm": {
+        "title": "Usunąć rodzaj opakowania?",
+        "message": "{{name}} zostanie usunięty. Produkty w nim zapakowane trzeba najpierw przepakować."
+      },
+      "validation": {
+        "non_negative": "musi wynosić zero lub więcej"
+      }
+    },
     "pages": {
       "settings": {
         "stock_locations": {
           "add_sheet_title": "Dodaj lokalizację magazynową",
           "edit_sheet_title": "Edytuj lokalizację magazynową",
           "search_placeholder": "Szukaj lokalizacji magazynowych…"
+        },
+        "package_types": {
+          "add_sheet_title": "Dodaj rodzaj opakowania"
         }
       },
       "products": {
